+++ drl_grasping/envs/models/__init__.py
@@ -1,3 +1,4 @@
+from .lights import *
 from .objects import *
 from .robots import *
 from .sensors import *