+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Configuration of pre-commit git hooks
 - Support for `DRL_GRASPING_BROADCAST_INTERACTIVE_GUI` environment variable.
 - Support for `DRL_GRASPING_SENSORS_RENDER_ENGINE` environment variable
+- Models for `Sun` and `RandomSun`, replacing the default light contained in SDF worlds.
 
 ### Changed